+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/Amebis/eduVPN/compare/3.255.19...HEAD)
 
 - Translations updated
+- Fixes: #241
 
 
 ## [3.255.19](https://github.com/Amebis/eduVPN/compare/3.255.18...3.255.19) (2024-05-03)