+++ CHANGES.md
@@ -6,7 +6,7 @@
 - Authorization time migration from 3.x client settings fixed
 - eduvpn-common updated to 4e7d42986b333147bbc89508c3ebeb9cc289374d and merged 7b6c4cce34fa34bba262f794c6702011f4edc0e3
 - Translations updated
-- Fixes: #241
+- Fixes: #241, #244
 
 
 ## [3.255.19](https://github.com/Amebis/eduVPN/compare/3.255.18...3.255.19) (2024-05-03)